speech/requirements.txt
paho-mqtt==1.3.1
google-cloud-speech==0.30.0
raven==6.0.0
